Add AdsCentral with AdPeriod and Categories creation

+++ KombfuscaWebManager/Models/AdModels/AdCategory.cs
@@ -6,6 +6,6 @@
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public int MaxAds { get; set; }
-        public int MinValue { get; set; }
+        public double MinValue { get; set; }
     }
 }
